pyproject.toml
[project]
name = "askimmigrate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chroma-hnswlib>=0.7.6",
    "chromadb>=0.6.3",
    "docx>=0.2.4",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-groq>=0.3.2",
    "langchain-huggingface>=0.3.0",
    "langchain-openai>=0.3.22",
    "langgraph>=0.4.8",
    "pymupdf>=1.26.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "python-slugify>=8.0.4",
    "pyyaml>=6.0.2",
    "streamlit>=1.45.1",
    "tiktoken>=0.9.0",
]
